requirements.txt
sqlalchemy
psycopg2
fastapi
uvicorn
python-dotenv
streamlit
pillow
requests